Reorganize routes names and backend organization
