+++ backend/models/permission.model.go
@@ -4,7 +4,7 @@ import "alexandrie/types"
 
 type Permission struct {
 	Id               types.Snowflake `json:"id" form:"id" binding:"omitempty"`
-	NodeId           types.Snowflake `json:"node_id" form:"node_id"`
+	NodeId           types.Snowflake `json:"node_id" form:"node_id" binding:"omitempty"`
 	UserId           types.Snowflake `json:"user_id" form:"user_id"`
 	Permission       int             `json:"permission" form:"permission"`
 	CreatedTimestamp int64           `json:"created_timestamp" form:"created_timestamp" binding:"omitempty"`
